Move type to types/index.ts file
