--- src/ts/Modifiers/Modifier.ts
@@ -1,3 +1,0 @@
-export default interface Modifier {
-    modify: () => void
-}
